Added D and started universal uint use

+++ cPrimes/cPrimes.c
@@ -11,21 +11,21 @@ back a structure instead of a tuple.
 #include <memory.h>
 #include <time.h>
 
-const long modPrimes[] = {7, 11, 13, 17, 19, 23, 29, 31};
-const long gaps[] = {4, 2, 4, 2, 4, 6, 2, 6, 4, 2, 4, 2, 4, 6, 2, 6};
-const long ndxs[] = {0, 0, 0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 7, 7, 7, 7, 7, 7};
+const unsigned int modPrimes[] = {7, 11, 13, 17, 19, 23, 29, 31};
+const unsigned int gaps[] = {4, 2, 4, 2, 4, 6, 2, 6, 4, 2, 4, 2, 4, 6, 2, 6};
+const unsigned int ndxs[] = {0, 0, 0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 6, 6, 7, 7, 7, 7, 7, 7};
 
 typedef struct {
-    long pc;
-    long maxprime;
+    unsigned int pc;
+    unsigned int maxprime;
 } primeResults;
 
-primeResults eratosthenes(long limit);
-primeResults iprimes2(long limit);
-primeResults primes235(long limit);
+primeResults eratosthenes(unsigned int limit);
+primeResults iprimes2(unsigned int limit);
+primeResults primes235(unsigned int limit);
 
 int main(int argc, char* argv[]) {
-    long limit;
+    unsigned int limit;
     primeResults eres, ip2, p235;
     double startTime, endTime, timeElapsed;
 
@@ -37,7 +37,7 @@ int main(int argc, char* argv[]) {
         limit = limit << (10-1);
     }
 
-    printf("Limit: %ld\n", limit);
+    printf("Limit: %d\n", limit);
 
     startTime = (float)clock()/CLOCKS_PER_SEC;
     eres = eratosthenes(limit);
@@ -45,8 +45,8 @@ int main(int argc, char* argv[]) {
     timeElapsed = endTime - startTime;
 
     printf("Eratosthenes: %.4fms\n", 1000*timeElapsed);
-    printf("\tPrimes counted: %ld\n", eres.pc);
-    printf("\tMax prime: %ld\n\n\n", eres.maxprime);
+    printf("\tPrimes counted: %d\n", eres.pc);
+    printf("\tMax prime: %d\n\n\n", eres.maxprime);
 
     startTime = (float)clock()/CLOCKS_PER_SEC;
     ip2 = iprimes2(limit);
@@ -54,8 +54,8 @@ int main(int argc, char* argv[]) {
     timeElapsed = endTime - startTime;
 
     printf("Iprimes2: %.4fms\n", 1000*timeElapsed);
-    printf("\tPrimes counted: %ld\n", ip2.pc);
-    printf("\tMax prime: %ld\n\n\n", ip2.maxprime);
+    printf("\tPrimes counted: %d\n", ip2.pc);
+    printf("\tMax prime: %d\n\n\n", ip2.maxprime);
 
     startTime = (float)clock()/CLOCKS_PER_SEC;
     p235 = primes235(limit);
@@ -63,14 +63,14 @@ int main(int argc, char* argv[]) {
     timeElapsed = endTime - startTime;
 
     printf("primes235: %.4fms\n", 1000*timeElapsed);
-    printf("\tPrimes counted: %ld\n", p235.pc);
-    printf("\tMax prime: %ld\n\n\n", p235.maxprime);
+    printf("\tPrimes counted: %d\n", p235.pc);
+    printf("\tMax prime: %d\n\n\n", p235.maxprime);
 
     return EXIT_SUCCESS;
 }
 
-primeResults eratosthenes(long limit) {
-    long i, j;
+primeResults eratosthenes(unsigned int limit) {
+    unsigned int i, j;
     primeResults results;
 
     results.pc = 4;
@@ -102,9 +102,9 @@ primeResults eratosthenes(long limit) {
     return(results);
 }
 
-primeResults iprimes2(long limit) {
-    volatile long pc, maxprime;
-    long lmtbf = (limit - 3) / 2;
+primeResults iprimes2(unsigned int limit) {
+    volatile unsigned int pc, maxprime;
+    unsigned int lmtbf = (limit - 3) / 2;
     primeResults results;
     results.pc = 0;
 
@@ -115,20 +115,20 @@ primeResults iprimes2(long limit) {
     char* buf = (char*)malloc((lmtbf+1)*sizeof(char));
     memset(buf,1,lmtbf+1);
     
-    for(long i = 0; i < ((long)sqrt(limit) - 3) / 2 + 1; i++) {
+    for(unsigned int i = 0; i < ((unsigned int)sqrt(limit) - 3) / 2 + 1; i++) {
         //printf("outeri: %ld\n", i);
         if(buf[i]) {
-            long p = i + i + 3;
-            long s = p * (i + 1) + i;
+            unsigned int p = i + i + 3;
+            unsigned int s = p * (i + 1) + i;
             //printf("p: %ld, s: %ld\n", p, s);
-            for(long j = s; j < (lmtbf + 1); j += p) {
+            for(unsigned int j = s; j < (lmtbf + 1); j += p) {
                 //printf("Setting %d j to off\n", j);
                 buf[j] = 0;
             }
         }
     }
 
-    for(long i = 0; i < lmtbf+1; i++) {
+    for(unsigned int i = 0; i < lmtbf+1; i++) {
         if(buf[i]) {
             pc++;
             maxprime = i + i + 3;
@@ -142,16 +142,17 @@ primeResults iprimes2(long limit) {
     return(results);
 }
 
-primeResults primes235(long limit) {
+primeResults primes235(unsigned int limit) {
     primeResults results;
     results.pc = 0;
     results.maxprime = 0;
-    volatile long pc = 0;
-    volatile long maxprime = 0;
-    long lmtbf = floor((limit + 23 ) / 30) * 8 - 1;
-    long lmtsqrt = (long)(sqrt(limit) - 7);
+    volatile unsigned int pc = 0;
+    volatile unsigned int maxprime = 0;
+    unsigned int lmtbf = floor((limit + 23 ) / 30) * 8 - 1;
+    unsigned int lmtsqrt = (unsigned int)(sqrt(limit) - 7);
+    //This needs to be signed!
     long lmtmod = lmtsqrt % 30;
-    long fIndex = 0;
+    unsigned int fIndex = 0;
     if(lmtmod < 0){
         fIndex = 30 + lmtmod;
     } else {
@@ -162,15 +163,15 @@ primeResults primes235(long limit) {
     memset(buf,1,lmtbf+1);
 
     //Algorithm
-    for(long i = 0; i < lmtsqrt+1; i++) {
+    for(unsigned int i = 0; i < lmtsqrt+1; i++) {
         if(buf[i]) {
-            long ci = i & 7;
-            long p = 30*(i>>3) + modPrimes[ci];
-            long s = p * p - 7;
-            long p8 = p << 3;
-            for(long j = 0; j < 8; j++) {
-                long c = floor(s/30)*8 + ndxs[s%30];
-                for(long jj = c; jj < lmtbf+1; jj += p8) {
+            unsigned int ci = i & 7;
+            unsigned int p = 30*(i>>3) + modPrimes[ci];
+            unsigned int s = p * p - 7;
+            unsigned int p8 = p << 3;
+            for(unsigned int j = 0; j < 8; j++) {
+                unsigned int c = floor(s/30)*8 + ndxs[s%30];
+                for(unsigned int jj = c; jj < lmtbf+1; jj += p8) {
                     buf[jj] = 0;
                 }
                 s += p * gaps[ci];
@@ -182,7 +183,7 @@ primeResults primes235(long limit) {
     pc = 3;
     maxprime = 7;
 
-    for(long i = 0; i < lmtbf-6+(ndxs[(limit-7)%30]); i++) {
+    for(unsigned int i = 0; i < lmtbf-6+(ndxs[(limit-7)%30]); i++) {
         if(buf[i]) {
             pc++;
             maxprime = 30*(i>>3) + modPrimes[i&7];
